Moved components folder away from pages folder
