+++ pages/companies/faglig.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Markdown from '../components/Markdown';
-import PageBody from '../components/MenuComponents/PageBody';
+import Markdown from '../../components/Markdown';
+import PageBody from '../../components/MenuComponents/PageBody'
 import Tabs from './Tabs';
 
 const InfoboxSource = `
