+++ pages/companies/info.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Markdown from '../components/Markdown';
+import Markdown from '../../components/Markdown';
 
 const InfoboxSource = `
   # Generell Info
