+++ pages/home/Homepage.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
-import PolygonSVG from '../components/FooterComponents/FooterPolygon.svg';
-import Menu from '../components/MenuComponents/Menu';
-import ThemeSwitcher from '../components/ThemeComponents/ThemeSwitchers';
-import LogoSVG from '../components/MenuComponents/Logo.svg';
+import PolygonSVG from '../../components/FooterComponents/FooterPolygon.svg';
+import Menu from '../../components/MenuComponents/Menu';
+import ThemeSwitcher from '../../components/ThemeComponents/ThemeSwitchers';
+import LogoSVG from '../../components/MenuComponents/Logo.svg';
 
 const LowerPolygon = styled(PolygonSVG)`
   fill: ${props => props.theme.orange};
